refactor(validator): validator 클래스를 NumbersValidator와 RetryValdiator 클래스로 분리
- 기존 Validator 클래스의 validate()로 재시작 유효성 검사를 하면 정답 숫자길이와 입력된 숫자 길이가 다르므로 IllegalArgumentException이 발생한다.
- 재시작 할 때는 중복 검사 체크를 하면 안된다
- 재시작 유효성 검사와 사용자 입력 숫자 유효성 검사를 나눠 재시작 숫자를 입력받을 시에는 중복 검사를 체크하지 않도록 하였다
