+++ src/main/java/baseball/domain/validator/RetryValidator.java
@@ -1,0 +1,14 @@
+package baseball.domain.validator;
+
+import java.util.regex.Pattern;
+
+public class RetryValidator extends Validator {
+
+    public static final Pattern RETRY_PATTERN = Pattern.compile("^1$|^2$");
+
+    @Override
+    public void validate(String input) {
+        super.validate(input);
+        checkRegularExpression(RETRY_PATTERN, input);
+    }
+}
